rename and move test, remove unneeded providers
